Relocate sport-only configuration
